% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_theme.R
\name{custom_theme}
\alias{custom_theme}
\title{A function set a custom ggplot theme.}
\usage{
custom_theme(base_size = 9)
}
\arguments{
\item{base_size}{Base size of plot text}
}
\description{
This function sets ggplot theme elements that I like, with the ability to change
the base size of the text.
}
\examples{
library(ggplot2)

ggplot(iris, aes(Sepal.Length, Sepal.Width)) +
    geom_point() +
    custom_theme(base_size = 10)

}
\keyword{plotting}
