% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahr_to_celsius.R
\name{fahr_to_celsius}
\alias{fahr_to_celsius}
\title{Convert temperature data from Fahrenheit to Celsius}
\usage{
fahr_to_celsius(fahr)
}
\arguments{
\item{fahr}{Temperature data in degrees Fahrenheit to be converted}
}
\value{
temperature value in degrees Celsius
}
\description{
Convert temperature data from Fahrenheit to Celsius
}
\examples{
fahr_to_celsius(32)
fahr_to_celsius(c(32, 212, 72))
}
\keyword{conversion}
